pyproject.toml
[tool.poetry]
name = "python-playwright-demo"
version = "0.1.0"
description = ""
authors = ["Konstantin Smyshliaev <smyshlyaevkonst@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.12"
pytest = "8.4.1"
allure-pytest = "2.15.0"
pydantic-settings = "2.10.1"
yarl = "1.20.1"
pytest-playwright = "0.7.0"
lxml = "6.0.0"


[tool.poetry.group.dev.dependencies]
black = "25.1.0"
isort = "6.0.1"
pre-commit = "4.2.0"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_format = "%(levelname)s %(asctime)s%(msecs)03d %(name)s: %(message)s"
log_date_format = "%d-%m-%Y %H:%M:%S:"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
